confirmacao de leitura

+++ Modules/Core/Services/UserService.php
@@ -9,6 +9,7 @@ use Exception;
 use Illuminate\Support\Facades\DB;
 use Modules\Core\Repositories\UserRepository;
 use Modules\Docs\Repositories\AgrupamentoUserDocumentoRepository;
+use Modules\Docs\Repositories\DocumentoRepository;
 use Modules\Docs\Repositories\UserEtapaDocumentoRepository;
 use Modules\Docs\Services\AgrupamentoUserDocumentoService;
 use Modules\Docs\Services\UserEtapaDocumentoService;
@@ -20,6 +21,7 @@ class UserService
     private $userRepository;
     protected $userEtapaDocumentoRepository;
     protected $agrupamentoUserDocumentoRepository;
+    protected $documentoRepository;
 
     public function __construct()
     {
@@ -28,6 +30,7 @@ class UserService
         $this->userRepository = new UserRepository();
         $this->userEtapaDocumentoRepository = new UserEtapaDocumentoRepository();
         $this->agrupamentoUserDocumentoRepository = new AgrupamentoUserDocumentoRepository();
+        $this->documentoRepository = new DocumentoRepository();
     }
 
     /** [
@@ -50,22 +53,27 @@ class UserService
                     ]
                 );
                 foreach ($buscaUserEtapaDocumento as $key => $value) {
-                    $arrayUserEtapaDocumento = [];
-                    $arrayUserEtapaDocumento['grupo_user_etapa'][$key] = [
-                        "user_id"           => (int) $data['user_substituto_id'],
-                        "grupo_id"          => $value->grupo_id,
-                        "etapa_fluxo_id"    => $value->etapa_fluxo_id,
-                    ];
-                    $arrayUserEtapaDocumento['documento_id'] = $value->documento_id;
-                    $arrayUserEtapaDocumento['documento_revisao'] = $value->documento_revisao;
 
-                    $userEtapaDocumentoService = new UserEtapaDocumentoService();
-                    if (!$userEtapaDocumentoService->delete((array) $value->id)['success']) {
-                        throw new Exception("Erro ao deletar o vinculo com o usuário.", 1);
-                    }
+                    $buscaDocumento = $this->documentoRepository->find($value->documento_id);
 
-                    if (!$userEtapaDocumentoService->store($arrayUserEtapaDocumento)['success']) {
-                        throw new Exception("Erro ao desvincular usuário.", 1);
+                    if ($buscaDocumento->revisao == $value->documento_revisao) {
+                        $arrayUserEtapaDocumento = [];
+                        $arrayUserEtapaDocumento['grupo_user_etapa'][$key] = [
+                            "user_id"           => (int) $data['user_substituto_id'],
+                            "grupo_id"          => $value->grupo_id,
+                            "etapa_fluxo_id"    => $value->etapa_fluxo_id,
+                        ];
+                        $arrayUserEtapaDocumento['documento_id'] = $value->documento_id;
+                        $arrayUserEtapaDocumento['documento_revisao'] = $buscaDocumento->revisao;
+
+                        $userEtapaDocumentoService = new UserEtapaDocumentoService();
+                        if (!$userEtapaDocumentoService->delete((array) $value->id)['success']) {
+                            throw new Exception("Erro ao deletar o vinculo com o usuário.", 1);
+                        }
+
+                        if (!$userEtapaDocumentoService->store($arrayUserEtapaDocumento)['success']) {
+                            throw new Exception("Erro ao desvincular usuário.", 1);
+                        }
                     }
                 }
 
@@ -75,26 +83,32 @@ class UserService
                     [
                         ['documento_id', '', $data['documentos'], "IN"],
                         ['user_id', '=', $data['user_id'], "AND"],
-                        ['grupo_id', '=', $data['grupo_id'], "AND"]
+                        ['grupo_id', '=', $data['grupo_id'], "AND"],
+                        ['documento_revisao', '=', $buscaDocumento->revisao, "AND"]
                     ]
                 );
                 foreach ($buscaGrupoTreinamentoDivulgacao as $key => $value) {
-                    $arrayAgrupamento = [];
-                    $arrayAgrupamento['grupo_and_user'][$key] = [
-                        "user_id"      => (int) $data['user_substituto_id'],
-                        "grupo_id"     => $value->grupo_id,
-                    ];
-                    $arrayAgrupamento['documento_id'] = $value->documento_id;
-                    $arrayAgrupamento['tipo'] = $value->tipo;
+                    $buscaDocumento = $this->documentoRepository->find($value->documento_id);
 
-                    $agrupamentoUserDocumentoService = new AgrupamentoUserDocumentoService();
+                    if ($buscaDocumento->revisao == $value->documento_revisao) {
+                        $arrayAgrupamento = [];
+                        $arrayAgrupamento['grupo_and_user'][$key] = [
+                            "user_id"      => (int) $data['user_substituto_id'],
+                            "grupo_id"     => $value->grupo_id,
+                            "tipo"         => $value->tipo
+                        ];
+                        $arrayAgrupamento['documento_id'] = $value->documento_id;
+                        $arrayAgrupamento['documento_revisao'] = $buscaDocumento->revisao;
 
-                    if (!$agrupamentoUserDocumentoService->delete((array) $value->id)['success']) {
-                        throw new Exception("Erro ao deletar o vinculo com o usuário.", 1);
-                    }
+                        $agrupamentoUserDocumentoService = new AgrupamentoUserDocumentoService();
 
-                    if (!$agrupamentoUserDocumentoService->store($arrayAgrupamento)['success']) {
-                        throw new Exception("Erro ao desvincular usuário.", 1);
+                        if (!$agrupamentoUserDocumentoService->delete((array) $value->id)['success']) {
+                            throw new Exception("Erro ao deletar o vinculo com o usuário.", 1);
+                        }
+
+                        if (!$agrupamentoUserDocumentoService->store($arrayAgrupamento)['success']) {
+                            throw new Exception("Erro ao desvincular usuário.", 1);
+                        }
                     }
                 }
             });
